+++ examples/ggui_examples/demo/fluids3d/box_data.py
@@ -1,4 +1,3 @@
-
 x_min = 0
 y_min = 0
 z_min = 0
@@ -8,5 +7,5 @@ y_max = 0
 z_max = 0
 
 verts = [
-    [x_min,x_min,x_min],
+    [x_min, x_min, x_min],
 ]
