+++ taichi/ui/backends/vulkan/canvas.h
@@ -12,6 +12,8 @@
 #include <array>
 #include <optional>
 #include <set>
+#include <memory>
+
 #include "taichi/ui/utils/utils.h"
 #include "taichi/ui/backends/vulkan/vertex.h"
 #include "taichi/ui/backends/vulkan/vulkan_utils.h"
@@ -20,15 +22,15 @@
 #include "taichi/ui/backends/vulkan/renderable.h"
 #include "taichi/ui/common/canvas_base.h"
 
-#include "gui.h"
-#include <memory>
+#include "taichi/ui/backends/vulkan/gui.h"
+#include "taichi/ui/backends/vulkan/renderer.h"
 
-#include "renderables/set_image.h"
-#include "renderables/triangles.h"
-#include "renderables/mesh.h"
-#include "renderables/particles.h"
-#include "renderables/circles.h"
-#include "renderables/lines.h"
+#include "taichi/ui/backends/vulkan/renderables/set_image.h"
+#include "taichi/ui/backends/vulkan/renderables/triangles.h"
+#include "taichi/ui/backends/vulkan/renderables/mesh.h"
+#include "taichi/ui/backends/vulkan/renderables/particles.h"
+#include "taichi/ui/backends/vulkan/renderables/circles.h"
+#include "taichi/ui/backends/vulkan/renderables/lines.h"
 
 TI_UI_NAMESPACE_BEGIN
 
@@ -36,9 +38,7 @@ namespace vulkan {
 
 class Canvas final : public CanvasBase {
  public:
-  Canvas(AppContext *app_context);
-
-  void prepare_for_next_frame();
+  Canvas(Renderer *renderer);
 
   virtual void set_background_color(const glm::vec3 &color) override;
 
@@ -50,44 +50,10 @@ class Canvas final : public CanvasBase {
 
   virtual void lines(const LinesInfo &info) override;
 
-  void mesh(const MeshInfo &info, Scene *scene);
-
-  void particles(const ParticlesInfo &info, Scene *scene);
-
   virtual void scene(SceneBase *scene_base) override;
 
-  void draw_frame(Gui *gui);
-
-  virtual void cleanup();
-
-  virtual void cleanup_swap_chain();
-
-  void recreate_swap_chain();
-
  private:
-  glm::vec3 background_color_ = glm::vec3(0.f, 0.f, 0.f);
-
-  std::vector<std::unique_ptr<Renderable>> renderables_;
-  int next_renderable_;
-
-  AppContext *app_context_;
-
-  VkSemaphore prev_draw_finished_vk_;
-  VkSemaphore this_draw_data_ready_vk_;
-
-  uint64_t prev_draw_finished_cuda_;
-  uint64_t this_draw_data_ready_cuda_;
-
-  std::vector<VkCommandBuffer> cached_command_buffers_;
-
- private:
-  void clear_command_buffer_cache();
-
-  void create_semaphores();
-  void import_semaphores();
-
-  template <typename T>
-  T *get_renderable_of_type();
+  Renderer *renderer_;
 };
 
 }  // namespace vulkan
